branches/main cisstRobot: Minor fix to compile in windows (robDOF.h, robSE3Track.h). CMN accessors and mutators added to robCollaborativeControlForce
git-svn-id: https://svn.lcsr.jhu.edu/cisst/branches/main@727 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ libs/include/cisstRobot/robSE3Track.h
@@ -30,7 +30,8 @@ namespace cisstRobot{
     real wmax;
     real wdmax;
 
-    uint32_t numwp;
+    //uint32_t numwp;
+    unsigned long int numwp;
 
     robError Evaluate(const robDOF& input);
